+++ history/backends/postgres.py
@@ -10,7 +10,8 @@ TRIGGER_FUNCTION_SQL = """
         _ctid integer := TG_ARGV[0]::integer;
         _pk_name text := TG_ARGV[1];
         _record_snap boolean := TG_ARGV[2]::boolean;
-        _fields text[] := TG_ARGV[3:];
+        _snap_of text := TG_ARGV[3];
+        _fields text[] := TG_ARGV[4:];
         _old jsonb := to_jsonb(OLD);
         _new jsonb := to_jsonb(NEW);
         _snapshot jsonb;
@@ -22,9 +23,15 @@ TRIGGER_FUNCTION_SQL = """
         END IF;
 
         IF _record_snap THEN
-            SELECT jsonb_object_agg(key, value) INTO _snapshot
-            FROM jsonb_each(_new)
-            WHERE key = ANY(_fields);
+            IF _snap_of = 'OLD' THEN
+                SELECT jsonb_object_agg(key, value) INTO _snapshot
+                FROM jsonb_each(_old)
+                WHERE key = ANY(_fields);
+            ELSEIF _snap_of = 'NEW' THEN
+                SELECT jsonb_object_agg(key, value) INTO _snapshot
+                FROM jsonb_each(_new)
+                WHERE key = ANY(_fields);
+            END IF;
         END IF;
 
         IF (TG_OP = 'UPDATE') THEN
@@ -133,17 +140,18 @@ class PostgresHistoryBackend(HistoryBackend):
             return tr_name, []
         self.execute(
             """
-                CREATE TRIGGER {tr_name} AFTER {trans_type} ON {table}
-                FOR EACH ROW EXECUTE PROCEDURE
-                history_record({ctid}, '{pk_col}', {snapshots}{field_list});
+            CREATE TRIGGER {tr_name} AFTER {trans_type} ON {table}
+            FOR EACH ROW EXECUTE PROCEDURE
+            history_record({ctid}, '{pk_col}', {snapshots}, '{snap_of}', {field_list});
             """.format(
                 tr_name=tr_name,
                 trans_type=trigger_type.name.upper(),
                 table=model._meta.db_table,
                 ctid=ct.pk,
                 pk_col=model._meta.pk.column,
-                snapshots=int(conf.SNAPSHOTS and trigger_type.snapshot),
-                field_list=", '" + "', '".join(field_names) + "'",
+                snapshots=int(conf.SNAPSHOTS),
+                snap_of=trigger_type.snapshot_of,
+                field_list="'" + "', '".join(field_names) + "'",
             )
         )
         return tr_name, field_names