+++ history/backends/sqlite.py
@@ -66,9 +66,9 @@ class SQLiteHistoryBackend(HistoryBackend):
         Returns an SQL fragment that builds a JSON object from the specified model
         fields.
         """
-        if not conf.SNAPSHOTS or not trigger_type.snapshot:
+        if not conf.SNAPSHOTS:
             return "NULL"
-        return self._json_object(fields, "NEW")
+        return self._json_object(fields, trigger_type.snapshot_of)
 
     def _json_changes(self, fields, trigger_type):
         """