+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.4.8
+
+* Removed `trigger_type.snapshot` so that delete triggers will now record snapshots.
+  Added `trigger_type.snapshot_of` to determine snapshot of OLD or NEW object.
+
+
 ## 3.4.7
 
 * Added a `HISTORY_IGNORE_MODELS` setting to ignore individual models. This should be a